+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Changed
 =======
 - Update endpoint ``GET v2/stored_flows`` to return the flows in order by ``priority`` descending and ``updated_at`` ascending.
 - Update endpoint ``GET v2/stored_flows`` to return the flows in descending order by `priority`.
+- When creating a flow, ``actions`` and ``instruction`` are not allowed to be present at the same time.
 - Reduced the number of flows shown in the logs. Maximum number of shown flows is 200.
 
 General Information